requirement.txt
uvicorn
python-dotenv
fastapi
loguru
langchain
langchain-community
langchain-openai
redis
werkzeug
pymysql
mysql-connector-python
streamlit
pyjwt
streamlit-toggle-switch
faiss-cpu
pdfplumber
python-docx
bs4
python-pptx
pytesseract
python-multipart
openpyxl
langchain-ollama
podcastfy